refactor(connector): extract token refresh to connector-service
Move refreshConnectorAccessToken from build-context.ts (private) to
connector-service.ts (exported) so it can be reused by the auth
endpoint for runtime OAuth token refresh.

upsertConnectorSecret is now internal to connector-service since its
only callers are within the same module.
